requirements.txt
argon2-cffi==16.3.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
configparser==3.5.0
dj-database-url==0.4.2
Django==1.11.7
django-appconf==1.0.2
django-contrib-comments==1.8.0
django-countries==4.6.2
django-filter==1.0.4
django-ipware==1.1.6
django-model-utils==3.0.0
django-notifications-hq==1.3
django-taggit==0.22.1
django-tracking-analyzer==0.2
django-tracking2==0.4.0
django-user-agents==0.3.2
djangorestframework==3.6.4
docutils==0.14
enum34==1.1.6
flake8==3.5.0
GeoIP==1.3.2
geoip2==2.5.0
gunicorn==19.7.1
idna==2.6
image==1.5.16
ipaddress==1.0.18
jsonfield==2.0.2
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
maxminddb==1.3.0
mccabe==0.6.1
oauthlib==2.0.4
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
ua-parser==0.7.3
urllib3==1.22
user-agents==1.1.0
whitenoise==3.3.1
